# src/swagger.json
{
  "basePath": "/",
  "consumes": ["application/json"],
  "info": {
    "title": "Passphrase Launchpad Server",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    }
  },
...
  "definitions": {
...
    "Invitations": {
      "properties": {
        "sender": {
          "type": "string",
          "ref": "User",
          "required": true
        },
        "recipient": {
          "type": "string",
          "lowercase": true,
          "trim": true,
          "required": true
        },
        "referralCode": {
          "type": "string",
          "required": true,
          "refPath": "User.referralCode",
          "default": null
        },
        "createdAt": {
          "type": "number",
          "default": null,
          "required": true
        }
      },
      "required": ["sender", "recipient", "referralCode", "createdAt"],
      "type": "object"
    }
  },
  "paths": {
...
    "/v1/invitations/search": {
      "get": {
        "tags": ["Invitations"],
        "operationId": "searchInvitations",
        "summary": "Search Invitations",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Invitations"
                  }
                },
                "count": {
                  "type": "number",
                  "default": 0
                },
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "searchInvitations",
              "type": "object",
              "required": ["success"]
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "recipient",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 0
          }
        ]
      }
    },
    "/v1/invitations": {
      "get": {
        "tags": ["Invitations"],
        "operationId": "getAllInvitations",
        "summary": "Get All Invitations",
        "description": "Endpoint for get all invitations",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/Invitations",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v2/invitations/search": {
      "get": {
        "tags": ["Invitations"],
        "operationId": "searchInvitations",
        "summary": "Search Invitations",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Invitations"
                  }
                },
                "count": {
                  "type": "number",
                  "default": 0
                },
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "searchInvitations",
              "type": "object",
              "required": ["success"]
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": []
      }
    },
    "/v2/invitations": {
      "get": {
        "tags": ["Invitations"],
        "operationId": "getAllInvitations",
        "summary": "Get All Invitations",
        "description": "Endpoint for get all invitations",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/Invitations",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/invitations/{id}": {
      "delete": {
        "tags": ["Invitations"],
        "operationId": "deleteInvitation",
        "summary": "Delete Invitation",
        "description": "Endpoint for delete invitation",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "boolean",
                  "default": true
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v2/invitations/{id}": {
      "delete": {
        "tags": ["Invitations"],
        "operationId": "deleteInvitation",
        "summary": "Delete Invitation",
        "description": "Endpoint for delete invitation",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "boolean",
                  "default": true
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/ondato_webhook/{identification_id}/receive": {
      "post": {
        "tags": ["Webhooks"],
        "operationId": "kycProcess",
        "summary": "KYC Process",
        "description": "Endpoint for KYC process",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully done",
            "schema": {
              "properties": {
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "identification_id",
            "schema": {
              "properties": {
                "identification_id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["identification_id"]
            }
          },
          {
            "in": "body",
            "name": "kycProcess",
            "schema": {
              "properties": {
                "Event": {
                  "type": "string",
                  "default": null
                },
                "RequestStatus": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["Event", "RequestStatus"]
            }
          }
        ]
      }
    },
    "/v2/ondato_webhook/{identification_id}/receive": {
      "post": {
        "tags": ["Webhooks"],
        "operationId": "kycProcess",
        "summary": "KYC Process",
        "description": "Endpoint for KYC process",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully done",
            "schema": {
              "properties": {
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "identification_id",
            "schema": {
              "properties": {
                "identification_id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["identification_id"]
            }
          },
          {
            "in": "body",
            "name": "kycProcess",
            "schema": {
              "properties": {
                "Event": {
                  "type": "string",
                  "default": null
                },
                "RequestStatus": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["Event", "RequestStatus"]
            }
          }
        ]
      }
    }
  },
  "produces": ["application/json"],
  "swagger": "2.0",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "body",
      "description": "This API requires authentication using a bearer token. Please include the token in the 'Authorization' header of your requests. ex: Bearer [space] <your-token>"
    }
  }
}